requirements.txt
instabot==0.117.0
openpyxl==3.0.7
osmnx==1.1.1
Pillow==8.3.1
twython==3.8.2
